fix(quickorder): child products instead of groups
There was a misconception so far: there is actually no need to focus on
product groups. Multiple products are added to the form using the id
prefix in the fields. That is sufficient to allow Foxy Cart to
understand the multiple products.
Also, groups of products are not the target of the subscription fields.
The whole form is.