refactor: move confirm dialog from private to internal
What's the difference? Private elements are supposed to be added to scoped element registries, but apparently using those can make it too complex for library consumers to customize our elements. Internal components, on the other hand, don't require scoping because they are automatically registered in the global scope with the `foxy-internal-` prefix. The goal is to make all "private" elements "internal" at some point.