fix(foxy-i18n): wrap shadow dom contents in a span
We often use `foxy-i18n` in buttons and links to provide translatable captions, and I've noticed that in iOS Safari that would make the entire area covered by text behave like it had `pointer-events: none` on it. In other words, if you used `foxy-i18n` in a button, users wouldn't be able to click that button. Seems that the internal Safari logic responsible for the "double tab" behaviour expects shadow roots to have a top-level element node to work properly, so I'm adding it here.