build: fix vaadin version
For some reason we've uploaded this project with Vaadin 15 listed in dependencies while officially we've always supported Vaadin 14 LTS only. Since there aren't any breaking changes between the two versions that'd impact Elements or its users, we are now correcting this mistake. Vaadin 14 LTS is a better option among the two because it will be supported for 2 more years – unlike v15 which is already deprecated.