refactor(foxy-nucleon): don't share `Rumour` updates on initial load
This behavior was useful in cases where some elements had stale versions of a resource and needed to be updated when a new version was loaded. However, it also introduced performance issues and the risk of update loops. Removing it in favor of more robust mechanisms like Nucleon’s local server for syncing resource state across the app.